// src/interfaces/ResourceList.ts
export default interface ResourceList {
  wood: number;
  component: number;
  water: number;
  alcheol: number;
  moonshine: number;
  cigarette: number;
  sugar: number;
  coffee: number;
  vegetable: number;
  rawFood: number;
  cannedFood: number;
  shovel: number;
  lockpick: number;
  filter: number;
  mechanicalPart: number;
  electricalPart: number;
  jewelry: number;
  book: number;
  brokenGuitar: number;
  guitar: number;
  weaponPart: number;
  brokenPistol: number;
  pistol: number;
  knife: number;
  shell: number;
  brokenAssaultRifle: number;
  assaultRifle: number;
  hatchet: number;
  ammo: number;
  brokenShotgun: number;
  shotgun: number;
  herb: number;
  chems: number;
  herbalMeds: number;
  meds: number;
  bandages: number;
}